Add new game objects and refactor Map class for better memory management
- New game objects: traps, buttons, stones, walls and exit door.
- Split map into map_loader and map_collision for better separation of concerns.

Others:
- Adjusted some small things.

+++ include/animation/collision.h
@@ -24,7 +24,7 @@ public:
 
     SDL_FRect getRect() const;
 
-    void move(const int& x, const int& y);
+    void move(const float& x, const float& y);
 
     void debugDraw();
